+++ Stripe/UIViewController+Stripe_KeyboardAvoiding.m
@@ -78,10 +78,8 @@ NS_ASSUME_NONNULL_BEGIN
     self.lastKeyboardFrame = keyboardFrame;
     
     if (self.managedScrollView) {
-        
         UIScrollView *scrollView = self.managedScrollView;
         UIView *scrollViewSuperView = self.managedScrollView.superview;
-        UIView *lastResponder = self.lastResponder;
         
         UIEdgeInsets contentInsets = scrollView.contentInset;
         UIEdgeInsets scrollIndicatorInsets = scrollView.scrollIndicatorInsets;
@@ -96,28 +94,11 @@ NS_ASSUME_NONNULL_BEGIN
         self.currentBottomInsetChange += bottomInsetDelta;
         scrollView.contentInset = contentInsets;
         scrollView.scrollIndicatorInsets = scrollIndicatorInsets;
-        
-        if (!lastResponder || bottomIntersection.size.height <= 0) {
-            scrollView.contentOffset = CGPointMake(0, -scrollView.contentInset.top);
-        }
-        else {
-            // the keyboard is visible
-            CGRect responderFrame = [lastResponder convertRect:lastResponder.bounds toView:scrollView];
-            CGRect keyboardFrameInScrollViewCoords = [scrollView convertRect:keyboardFrame fromView:nil];
-            CGPoint offset = scrollView.contentOffset;
-            
-            CGFloat topOfScreenOffset = CGRectGetMinY(responderFrame);
-            CGFloat topOfKeyboardOffset = CGRectGetMinY(responderFrame) - CGRectGetMinY(keyboardFrameInScrollViewCoords);
-            offset.y = ((topOfScreenOffset + topOfKeyboardOffset) / 2) - scrollView.contentInset.top;
-            offset.y = MAX(offset.y, -scrollView.contentInset.top);
-            scrollView.contentOffset = offset;
-        }
     }
     
     if (self.keyboardFrameBlock) {
         self.keyboardFrameBlock(keyboardFrame, self.lastResponder);
     }
-    
 }
 
 @end